requirements.txt
appdirs==1.4.4
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
decorator==4.4.2
Django==3.1.5
django-cleanup==5.1.0
django-crispy-forms==1.10.0
django-etc==1.3.0
django-filter==2.4.0
django-hitcount==1.3.3
django-taggit==1.3.0
djangorestframework==3.12.2
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
iniconfig==1.1.1
moviepy==1.0.3
mypy-extensions==0.4.3
numpy==1.19.5
packaging==20.8
pathspec==0.8.1
Pillow==8.1.0
pluggy==0.13.1
proglog==0.1.9
py==1.10.0
pyparsing==2.4.7
pytest==6.2.1
pytest-django==4.1.0
pytest-pythonpath==0.7.3
pytz==2020.5
regex==2020.11.13
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
tqdm==4.56.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.2
